Move things around a bit more
