+++ npm/src/lib/semver.js
@@ -1,6 +1,7 @@
 // @flow
 
 import * as semver from "semver";
+import request from "request";
 
 type VersionRange = ">=" | "<=";
 export type Version = {
